:recycle: Move Timestamp to toolbox
