+++ src/components/toolbox/timestamp/index.js
@@ -3,7 +3,7 @@ import moment from 'moment';
 import Tooltip from 'react-toolbox/lib/tooltip';
 import theme from 'react-toolbox/lib/tooltip/theme.css';
 import { translate } from 'react-i18next';
-import i18n from '../../i18n';
+import i18n from '../../../i18n';
 
 const _convertTimeFromFirstBlock = value =>
   new Date((((Date.UTC(2016, 4, 24, 17, 0, 0, 0) / 1000) + value) * 1000));
